Refactored module config
