+++ src/main/java/frc/robot/competition/CompetitionRobotContainer.java
@@ -22,6 +22,7 @@ import edu.wpi.first.wpilibj2.command.button.RobotModeTriggers;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine.Direction;
 import frc.robot.classes.BaseDrive;
 import frc.robot.classes.ModuleConfig;
+import frc.robot.classes.Structs.ClosedLoopParameters;
 import frc.robot.commands.FieldOrientedDrive;
 import frc.robot.commands.FieldOrientedWithCardinal;
 import frc.robot.commands.OrbitalTarget;
@@ -173,14 +174,13 @@ class CompetitionRobotContainer {
   }
 
   private NeoModule makeSwerveModule(int driveId, int steerId) {
-    ModuleConfig.ClosedLoopParameters driveClosedLoopParams =
-        new ModuleConfig.ClosedLoopParameters(0.1, 0, 0, 1 / RobotConstants.DRIVE_WHEEL_FREESPEED);
-    ModuleConfig.ClosedLoopParameters steerClosedLoopParams =
-        new ModuleConfig.ClosedLoopParameters(18, 0, 0, 0);
+    ClosedLoopParameters driveClosedLoopParams =
+        new ClosedLoopParameters(0.1, 0, 0, 1 / RobotConstants.DRIVE_WHEEL_FREESPEED);
+    ClosedLoopParameters steerClosedLoopParams = new ClosedLoopParameters(18, 0, 0, 0);
     return new NeoModule(
+        driveId,
+        steerId,
         new ModuleConfig(
-            driveId,
-            steerId,
             driveClosedLoopParams,
             steerClosedLoopParams,
             RobotConstants.DRIVE_ENC_TO_METERS,
